be/socketController: phase2 / art vote on, off

See merge request s07-blockchain-nft-sub2/S07P22A509!59

+++ server/src/main/java/com/classic/imteller/api/dto/room/EffectDto.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class EffectDto {
     private long uid;
-    private long effect;
+    private int effect;
     private int effectNum;
 }
